Add flake8 lint job
So, this project uses `black` and will until the heat death of
the universe, but we also need a way to learn about unused imports
and things of that sort.

So disable flake8 rules that are not compatible with black and treat
flake8 as a linter for those kinds of errors in our code.

Related-to: TCCP-831

Signed-off-by: Leonardo Held <leonardo.held@toradex.com>