Solve packaging issue with pid_map.yaml
This file should be ingested not from the current dir but from the absolute
path of the Aval container.

This wasn't caught in CI because the workdir from the container will be
different than the one from using it outside Aval's CI context. Forcing the
e2e test jobs to NOT clone the repo setting GIT_STRATEGY to 'none' will
prevent regressions.

Signed-off-by: Leonardo Held <leonardo.held@toradex.com>