Add a fail-fast mode to device locking mechanism
This commit makes Aval try to rapidly lock devices without waiting for a
timeout, meaning if a device is locked, Aval tries the next one instead of
waiting.

If all of the matched devices are locked, try to lock the last one parsed.

The use-case here is that you might want to add more devices of the same type
to speed up testing.

Was this a bug? I have no idea.

Related-to: AVAL-4

Signed-off-by: Leonardo Held <leonardo.held@toradex.com>