Check if module updated before remove fuse secondary
Previously, a hardcoded sleep was used, and AVAL trusted 180 seconds to update
the module. However, if the module takes more than 180 seconds to update, the
secondary fuse was removed from the initial version instead of the updated
version. This commit checks if the module is updated before trying to remove
the fuse.

Related-to: AVAL-59

Signed-off-by: Lucas Bernardes <lucas.bernardes@toradex.com>