Remove hardcoded paths for pid_map and main.py
The previously hardcoded paths only worked when running AVAL inside a
container, where the code is located at /aval. However, when running AVAL in a
regular shell, those paths become invalid as they incorrectly point to the
hardcoded "/aval" location, which may not reflect the actual path where AVAL
is located.

This commit replaces the hardcoded paths with relative paths, allowing AVAL to
work both inside containers and when executed from a shell.

It was necessary to add a "cd /aval" before executing AVAL's entrypoint because
GitLab CI clones the repository inside the container at the "/builds/..." path
and sets that directory as the job's default WORKDIR. This overrides the
WORKDIR defined in the Dockerfile of the image specified in "image:".

Related-to: TIE-1042

Signed-off-by: Lucas Bernardes <lucas.bernardes@toradex.com>