e2e-tests: Add tests for Shell and PowerShell
These tests expand the scope of the end-to-end (e2e) test coverage, that was
brought from the AVAL implementation within the Torizon IDE extension in native
shell and PowerShell (pwsh) environments. The goal of this commit is to ensure
AVAL works properly when run natively in both shell and pwsh contexts, beyond
containerized execution.

Related-to: TIE-1042

Signed-off-by: Lucas Bernardes <lucas.bernardes@toradex.com>