e2e-tests: Fix runner tag in jobs to access correct runners
Commit de70b534e28dbb4d6a177e5a4992ed5a35cef727 aimed to define tags in
jobs that use internal resources. This commit defined the tag "docker"
in this jobs, however the correct tag is "all".
Also, use "pwsh" instead of "vstudio" tag, because this tag is more
descritive about what runner are being used.
Finally, it was requested that AVAL uses "aws-toradex-large" runner tag
in jobs that doesn't require internal resources.

Signed-off-by: Lucas Bernardes <lucas.bernardes@toradex.com>