gitlab-ci: Run only 'sentry-watch' job in Sentry pipelines
It's not needed to build Sentry container in every Sentry pipeline, it's
possible to build this when a new push is added in 'main' branch, and
only run the execution in sentry pipelines. The built container will
only change if the code changed.
Additionaly, it's not needes to run flake8-lint and validate-yaml in
Sentry pipeline, it's enough do run this jobs in developing pipelines.

Signed-off-by: Lucas Bernardes <lucas.bernardes@toradex.com>