e2e-tests: Implement "pull_policy: always" to always pull containers
Horw runners use by default "pull_policy=if-not-present". Then the jobs only
performs pull of the container in “image:” field if it is not already available
locally on the runner. However, this is problematic for projects like AVAL,
which build their containers with the main tag and make them available to other
projects. This can result in the version of AVAL being used not being the
latest one, undermining the reliability of what is being tested.

Related-to: AVAL-79

Signed-off-by: Lucas Bernardes <lucas.bernardes@toradex.com>