gitlab-ci: Use saas-linux-small-amd64 as the default cloud runner
This change to the GitLab runners is a request from the automation team, which
is now adopting the GitLab runners as the default runners to be used by the
project.

A study was carried out to determine which GitLab runner is the best fit for
AVAL. In this study, a full AVAL pipeline was executed using each of GitLab’s
saas-linux runners, from small up to 2xlarge. The study showed that there was
little performance gain among these runners for AVAL. Therefore, it was
decided to use the small runner, since it has the lowest Cost Factor among
them.

Related-to: AVAL-80

Signed-off-by: Lucas Bernardes <lucas.bernardes@toradex.com>