sentry: Run sentry-watch also in push pipelines
Previously, there was an issue where the sentry-watch job was not tested in
push pipelines, which occasionally led to unexpected errors being merged into
the main branch. These errors were only discovered once the job actually ran
on the main branch.

This commit ensures that the job runs regardless of whether the WATCH variable
is defined, enabling proper testing in pushs of this job. The behavior of
running only this job in scheduled pipelines every 15 minutes remains valid.

Additionally, remove the "rules:" definition from the sentry-watch job, as it
is unnecessary — these rules are already defined through "<<: *run_on_watch"
and "<<: *skip_on_watch".

Signed-off-by: Lucas Bernardes <lucas.bernardes@toradex.com>