common: Correctly handle "hardware_id" of smarc devices
These devices has 2 hyphens in its "hardware_id", but it was being parsed as
if it were one of the devices that have only 1 hyphen in their name. The smarc
devices follow the same "hardware_id" pattern as devices that have "emmc" in
their "hardware_id". For this reason, it should be parsed in a similar way to
the devices with "emmc" in the "hardware_id".

Related-to: AVAL-83

Signed-off-by: Lucas Bernardes <lucas.bernardes@toradex.com>