common: Fix "parse_hardware_id()" function returning incorrect value for x86
In Cloud, the "deviceId" of the x86 devices follows the pattern
"torizon-x86-<id>", but the hardwareId is "intel-corei7-64", which does not
match the convention adopted in Torizon, where the hardwareId is usually
contained within the "deviceId". For this reason, a specific case for x86 is
created in the parse_hardware_id function.

Related-to: AVAL-88

Signed-off-by: Lucas Bernardes <lucas.bernardes@toradex.com>